+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [server] Save last user of deploy
 - [server] Fix long running exec commands
 - [server] Fix lack of error handling when streaming messages from readers
+- [client] Add limits validation on app creation
 
 ## [0.22.0] - 2018-06-07
 ### Changed